+++ Client/src/components/language-selector/index.js
@@ -219,10 +219,10 @@ class LanguageSelector extends React.PureComponent {
                                         className={"langDropdown"} 
                                         path={pageContext && pageContext.originalPath
                                             ? item.value === "en" && "" + (pageContext.originalPath === "/" ? "/" : pageContext.originalPath.replace(/(\/)?$/, ''))
-                                            : item.value === "en" && "/"}  
+                                            : item.value === "en" && ""}  
                                         href={pageContext && pageContext.originalPath
-                                            ? item.value !== "en" && "/" + item.value + (pageContext.originalPath === "/" ? "/" : pageContext.originalPath.replace(/(\/)?$/, ''))
-                                            : item.value !== "en" && "/" + item.value}
+                                            ? item.value !== "en" && "/" + item.value + (pageContext.originalPath === "/" ? "" : pageContext.originalPath.replace(/(\/)?$/, ''))
+                                            : item.value !== "en" && "" + item.value}
                                         headerText={item.text} 
                                         langChange={item.value === "en"}
                                     />
@@ -233,10 +233,10 @@ class LanguageSelector extends React.PureComponent {
                                         headerTextClass="selected" 
                                         path={pageContext && pageContext.originalPath 
                                             ? item.value === "en" && "" + (pageContext.originalPath === "/" ? "/" : pageContext.originalPath.replace(/(\/)?$/, ''))
-                                            : item.value === "en" && "/"}  
+                                            : item.value === "en" && ""}  
                                         href={pageContext && pageContext.originalPath
-                                            ? item.value !== "en" && "/" + item.value + (pageContext.originalPath === "/" ? "/" : pageContext.originalPath.replace(/(\/)?$/, ''))
-                                            : item.value !== "en" && "/" + item.value}
+                                            ? item.value !== "en" && "/" + item.value + (pageContext.originalPath === "/" ? "" : pageContext.originalPath.replace(/(\/)?$/, ''))
+                                            : item.value !== "en" && "" + item.value}
                                         langChange={item.value === "en"}
                                         headerText={item.text} />
                             )}
@@ -252,11 +252,11 @@ class LanguageSelector extends React.PureComponent {
                                         className={"langDropdown " + item.value} 
                                         path={pageContext && pageContext.originalPath
                                             ? item.value === "en" && "" + (pageContext.originalPath === "/" ? "/" : pageContext.originalPath.replace(/(\/)?$/, ''))
-                                            : item.value === "en" && "/" }
+                                            : item.value === "en" && "" }
                                         href={
                                             pageContext && pageContext.originalPath
-                                                ? item.value !== "en" && "/" + item.value + (pageContext.originalPath === "/" ? "/" : pageContext.originalPath.replace(/(\/)?$/, ''))
-                                                : item.value !== "en" && "/" + item.value
+                                                ? item.value !== "en" && "/" + item.value + (pageContext.originalPath === "/" ? "" : pageContext.originalPath.replace(/(\/)?$/, ''))
+                                                : item.value !== "en" && "" + item.value
                                         }
                                         langChange={item.value === "en"}
                                         headerText={item.text}